+++ imagestorage/src/main/java/com/woowacourse/imagestorage/strategy/convert/StaticImg2WebpStrategy.java
@@ -1,17 +1,17 @@
 package com.woowacourse.imagestorage.strategy.convert;
 
-import com.sksamuel.scrimage.ImmutableImage;
-import com.sksamuel.scrimage.webp.WebpWriter;
 import com.woowacourse.imagestorage.exception.FileConvertException;
+import com.woowacourse.imagestorage.strategy.convert.handler.CwebpHandler;
 import java.io.IOException;
 
 public class StaticImg2WebpStrategy implements Convert2WebpStrategy {
+
+    private static final CwebpHandler DEFAULT = new CwebpHandler();
+
     @Override
     public byte[] convert(final byte[] originBytes) {
         try {
-            return ImmutableImage.loader()
-                    .fromBytes(originBytes)
-                    .bytes(WebpWriter.DEFAULT);
+            return DEFAULT.convert(originBytes);
         } catch (IOException exception) {
             throw new FileConvertException("webp 변환 시 문제가 발생하였습니다.");
         }
